Changed the math function template.
The called functions are cdecl, thus the macro is responsible for cleaning up the stack.
There still seems to be a stack corruption I can't make sense of. It goes away if the function is disallowed from being inlined, but without understanding why this is, this patch should be looked at with suspicion.