Added PointerExpiredNotification.
If a class creates an object of a type it usually wouldn't usually receive an invalid pointer notification, the class can subscribe to these notifications using this class.

For example, if the ParasiteClass creates a squid grapple animation, it would not be notified when it is destroyed. Thus, it has to subscribe to NotifyInvalidAnim. When the grapple animation is destroyed (either by the ParasiteClass or in some external code), the ParasiteClass then unsubscribes from the NotifyInvalidAnim notifications.