Made the new host and hook sections optional.
Define /DSYR_VER=2 to enable the new behavior. If this is not done, two empty macros are defined thus code does not have to be updated to support both versions at the same time. (DEFINE_HOOK_AGAIN still has to be moved out between DEFINE_HOOK and the respective method body.)

Because Syringe.h is #included in many different files and SYR_VER has to be defined before this happens, it usually is unsafe to just #define SYR_VER 2 somewhere in code. This is because it is easy to miss some compilation units and thus not all hooks might be defined using the new version.