Fixed return values for OwnedTiberiumStruct
AddAmount and RemoveAmount both return a float. This cannot be neglected like with functions returning an integral value in a register, because it unbalances the floating point stack. Without a return value, the compiler doesn't add a fstp instruction to discard the float the function adds. This leads to freezes in unrelated places.