Updates to BombClass and BombListClass
Renamed OwnerObject to Owner, because Owner is a Techno, not an Object. Renamed and re-typed TargetUnit, because it could be an Object. (ObjectClasses don't update the bomb, so even though they can be attached to them, they will just tick forever).

Made the constructor public, but added a note about BombListClass::Plant.

Renamed the UpdateDelay member. It counts down, and if it reaches 0, the Detectors' BombSight is refreshed.