Added overloads for AbstractClass::GetCoords that return CoordStructs directly.
The virtual methods will stay, because they are the de-facto representation of the functions in the executable.

This is not as efficient as calling GetCoords with a pointer to CoordStruct which is then assigned in-place. This function might need to copy the return value to its final location. This is what the game does, so it is at least not slower as the game's original code.