Replaced the memory allocator.
operator new and operator delete are available in the YRMemory namespace. The allocator GameAllocator that calls the YR memory functions has been added. It can be used wherever an allocator is required.

The class Memory contains functions to allocate and construct objects from a given allocator, or arrays of constructed objects. Both support a variable number of arguments, and the constructor is chosen appropriately. For arrays, each element is constructed using the given arguments. This can also be an instance of the type to create, if the copy constructor is available.

The free functions [Game|DLL][Create|Delete][|Array] can be used like the previous macros, and they remove the need to pass an allocator.

Note that the array deallocating functions need to know the size of the array (elements, not bytes) for destroying that number of elements. This is required because no data is allocated to store such information.

The commented out macros show how the new functions are called using the old notation.