Merged dupe definitions.
SessionOptionsClass equaled GameModeOptionsClass. Kept and updated the latter.

StartAISlots equaled AISlotsStruct. Kept the definition in GameModeOptionsClass.h, but renamed it.

Also added a dummy member to account for all the other undefined fields in SessionClass, which also contains the Message List and hundreds of other fields.

TODO: This means that GameModeOptionsClass::Instance is the same as SessionClass::Instance->Config. This should be resolved some day. Also, lots of the references in Unsorted.h can be accessed using SessionClass.