Array classes reworked
* Removed unused includes, reordered methods, made stuff more consistent.

* IsEqual() converted to operator ==, which seems more appropriate. Added operator !=, calling operator ==.

* Support for move semantics. Vectors can be moved and swapped now. Items are tried to be moved when adding or resizing.

* Removed Load() and Save() methods. They will not work like this anyhow.

* Added the missing methods to CounterClass, made some of them const for read-only access.

* Added a static_assert for types that will not work as fast as they should because of alignment. If an int is preceding the actual block of memory for types requiring 8 byte alignment or higher, they will be shifted by 4 bytes. If that happens, better support for this has to be added.

* Removed some checks from begin() and end(), because they should only fire when the vector has not been constructed yet (in which case there's a worse problem somewhere), or when the vector is iterated while being resized, which generally is a bad idea.

Destructors:
As the VectorClass destructor is virtual, it is invoked for all derived classes on destruction. When this happens, calling Clear() calls VectorClass::Clear() (opposed to a overridden Clear()) because the vtable is reset when entering the base destructor. Thus, derived classes do not need to manually call the base class Clear().

Ideas for later:
operator == could use std::equal, which would take care of length and item comparison. Likewise, std::find and std::copy_n could be used in various places, and maybe std::rotate when removing an item.

When available, methods should be marked as nothrow. The original implementation would never throw an exception.