A better FacingStruct implementation
Fully implemented instead of jumping to game methods, which should help the compiler to optimize a few calls. All (?) missing functions are implemented.

No more pointer in pointer out functions, which most likely was the result of returning user defined types in the original source.

No more hidden features like the undiscoverable cast to int meaning map to 0..7.

The entire thing is encapsulated, which is a nice design though the rest of the code still pokes everything it needs.

Notes on the implementation:
ROT is still saved in a DirStruct, which might not be what the original code does. Maybe there is a common underlying struct, like a fixed point decimal value.