Implemented the Checksummer, removed calls of game methods
This implementation has no out of bounds writes beyond the class bounds. See the added check in the Fill method for ByteIndex == Size.

Several functions are split into a public method and an internal inline function. This is to prevent lots of function calls in methods that are called thousands of times potentially.

The Checksummer now accepts all values without casting. There is a function template that takes care of all types, and calls a single processing function, so not every type gets its own instantiation of the big function.

The current value can be retrieved using GetValue(), which fills the current buffer and also considers its value. Immediate() returns the raw value not considering the current buffer. Commit() fills the current buffer, processes and then clears it. Calling GetValue() before Commit() returns the same value as Immediate() after Commit(). GetValue() doesn't change the internal state, but it is not bitwise const as it fills the buffer.