AbstractTypeClass::FindOrAllocate now actually checks for 'none' and '<none>'.
It was possible FindOrAllocate was called with this special values, and then would use them to create type instances using these IDs.

Apparently, the game creates a HouseTypeClass with ID '<none>'. This could break FindIndex, because if '<none>' isn't found, its index would be -1, which it isn't if there is an item with that ID.