RadioCommand enum class
Replaced the long list of unnamed and untyped defines with a long list of named and typed values with comments. The comments are coming from the earliest C&C games with almost no changes.

4 to 6 are yet unknown, and a quick glance into earlier games didn't reveal anything. Some seem to be unused, like 16, 26 and 27.

For the naming: It's generally unintuitive. Values beginning with Answer are return values, all others are commands. Notify commands always return the same value, and thus are fire-and-forget. Query return a status answer, and Requests order the recipient to do something and return success or failure codes (Notify can also do this, but can't fail). A mess.

Also updated all functions using RadioCommand as parameter or return value.