Enter noinit_t
noinit_t is a unique type that can be used to create special constructors that do not initialize the class. This replaces the abuse of the default constructor or the protected constructors taking a bool.

Constructors taking noinit_t are optimal in terms of doing no initialization, thus it is recommended to delegate to the noinit constructor even in constructors that jump into game code. Goal is that each member is constructed using a noinit constructor and thus everything is inlined into an empty function.