Changed the definition of FindObjectNearestTo/FindTechnoNearestTo
The function only checks TechnoClass derivatives and thus only returns TechnoClass pointers. Thus, the excluded pointer has to be a TechnoClass, otherwise it doesn't compare equal to anything the method checks. Could have stayed ObjectClass, but why give the impression to the caller that this would have an effect?

The first param apparently is a pixel offset into the cell coords, which is used, but actually has no effect whatsoever. This is because the initial best match value is 0, and no distance from the offset can be less than that, the method actually only traverses the entire linked list and only ever returns the first TechnoClass content it finds.

I'm keeping the name with Find to indicate this method is kinda needlessly expensive (linked list traversal, calculating distance using square roots) and nothing like a cheap getter.