Changed how out of memory is handled
Previously, GameAllocator just returned a pointer that was maybe null. The functions using it were responsible to check. GameCreate and friends did, and terminated if a null pointer was returned.

Now the GameAllocator calls a new helper method that checks whether a null pointer has been returned and reports this by terminating the process. This means all methods later in the chain can assume they got a valid pointer, because otherwise they would not get control back.